correction Sauvegarde + Test

+++ Sources/test/ollert/SauvegardeTest.java
@@ -23,29 +23,20 @@ public class SauvegardeTest {
         page3.obtenirListe(0).creerTache("Tache 01");
         page3.obtenirListe(0).creerTache("Tache 02");
         page3.obtenirListe(1).creerTache("Tache 11");
-
-
-        File dossier = new File("ressource");
-        if (dossier.exists() && dossier.isDirectory()) {
-            // Liste des fichiers dans le dossier
-            File[] fichiers = dossier.listFiles();
-            if (fichiers != null) {
-                for (File fichier : fichiers) {
-                    fichier.delete();
-                }
-            }
-        }
     }
 
-    @AfterAll
-    public static void close(){
-        File dossier = new File("ressource");
+    @AfterEach
+    public void afterUp(){
+        File dossier = new File(Sauvegarde.DIR);
         if (dossier.exists() && dossier.isDirectory()) {
             // Liste des fichiers dans le dossier
             File[] fichiers = dossier.listFiles();
             if (fichiers != null) {
                 for (File fichier : fichiers) {
-                    fichier.delete();
+
+                    // si fichier .ol => suppression
+                    if (fichier.getName().endsWith(".ol"))
+                        fichier.delete();
                 }
             }
         }
@@ -55,16 +46,12 @@ public class SauvegardeTest {
     public void test_sauvegarderPage(){
         // sauvegarde valide
         assertTrue(Sauvegarde.sauvegarderPage(page1));
-        File fichier1 = new File("ressource/Ollert1.ol");
+        File fichier1 = new File(Sauvegarde.DIR+"Ollert1.ol");
         assertTrue(fichier1.exists());
 
         // verifications fichier non présent (normal)
-        File fichier2 = new File("ressource/Ollert2.ol");
+        File fichier2 = new File(Sauvegarde.DIR+"Ollert2.ol");
         assertFalse(fichier2.exists());
-
-        // verification non présence fichier mauvaise place
-        File fichier3 = new File("Ollert1.ol");
-        assertFalse(fichier3.exists());
     }
 
     @Test
@@ -75,18 +62,18 @@ public class SauvegardeTest {
 
         // fichier valide
 
-        Page page1Chargee = Sauvegarde.chargerPage("ressource/Ollert1.ol");
+        Page page1Chargee = Sauvegarde.chargerPage("Ollert1.ol");
         assertEquals(page1.toString(), page1Chargee.toString());
 
-        Page page2Chargee = Sauvegarde.chargerPage("ressource/Ollert2.ol");
+        Page page2Chargee = Sauvegarde.chargerPage("Ollert2.ol");
         assertEquals(page2.toString(), page2Chargee.toString());
 
-        Page page3Chargee = Sauvegarde.chargerPage("ressource/Ollert3.ol");
+        Page page3Chargee = Sauvegarde.chargerPage("Ollert3.ol");
         assertEquals(page3.toString(), page3Chargee.toString());
 
 
         // fichier non valide => null
-        Page page4Chargee = Sauvegarde.chargerPage("ressource/Ollert4.ol");
+        Page page4Chargee = Sauvegarde.chargerPage("Ollert4.ol");
         assertNull(page4Chargee);
     }
 }
